requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Enhanced Scanner Packages
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
pdf2image==1.16.3
easyocr==1.7.0
pymupdf==1.23.8